refactor : 패키지 분류, 객체 네이밍 변경
