+++ src/main/java/com/yapp/archiveServer/archive/domain/HomeCategoryId.java
@@ -1,7 +1,6 @@
-package com.yapp.archiveServer.homecategory.domain;
+package com.yapp.archiveServer.archive.domain;
 
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
+import lombok.*;
 
 import javax.persistence.Embeddable;
 import java.io.Serializable;
@@ -9,6 +8,8 @@ import java.io.Serializable;
 @Getter
 @Embeddable
 @EqualsAndHashCode
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class HomeCategoryId implements Serializable {
 
     private Long categoryId;
